+++ .gitignore
@@ -1,4 +1,3 @@
-.rubocop-*
 /.bundle/
 /.yardoc
 /Gemfile.lock
@@ -9,8 +8,10 @@
 /spec/reports/
 /tmp/
 
+.rubocop-*
+/bundler.d/*
+!/bundler.d/.keep
 /config/settings.local.yml
 /config/settings/*.local.yml
 /config/environments/*.local.yml
-
 /spec/manageiq
