+++ .rubocop_cc.yml
@@ -1,5 +1,4 @@
 inherit_from:
-# this is downloaded by .codeclimate.yml
 - .rubocop_base.yml
 - .rubocop_cc_base.yml
 - .rubocop_local.yml
