+++ .travis.yml
@@ -1,17 +1,9 @@
----
-dist: xenial
 language: ruby
+cache: bundler
 rvm:
 - 2.5.7
 - 2.6.5
-sudo: false
-cache: bundler
-env:
-  global:
-  - RUBY_GC_HEAP_GROWTH_MAX_SLOTS=300000
-  - RUBY_GC_HEAP_INIT_SLOTS=600000
-  - RUBY_GC_HEAP_GROWTH_FACTOR=1.25
 addons:
   postgresql: '10'
 install: bin/setup
-after_script: bundle exec codeclimate-test-reporter
+after_script: bin/ci/after_script
