+++ README.md
@@ -1,19 +1,17 @@
-# manageiq-providers-azure_stack
+# ManageIQ::Providers::AzureStack
 
-[![Gem Version](https://badge.fury.io/rb/manageiq-providers-azure_stack.svg)](http://badge.fury.io/rb/manageiq-providers-azure_stack)
 [![Build Status](https://travis-ci.com/ManageIQ/manageiq-providers-azure_stack.svg?branch=master)](https://travis-ci.com/ManageIQ/manageiq-providers-azure_stack)
-[![Code Climate](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack.svg)](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack)
-[![Test Coverage](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack/badges/coverage.svg)](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack/coverage)
+[![Maintainability](https://api.codeclimate.com/v1/badges/e9abcd437efff98ed5b9/maintainability)](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack/maintainability)
+[![Test Coverage](https://api.codeclimate.com/v1/badges/e9abcd437efff98ed5b9/test_coverage)](https://codeclimate.com/github/ManageIQ/manageiq-providers-azure_stack/test_coverage)
 [![Security](https://hakiri.io/github/ManageIQ/manageiq-providers-azure_stack/master.svg)](https://hakiri.io/github/ManageIQ/manageiq-providers-azure_stack/master)
 
 [![Chat](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/ManageIQ/manageiq-providers-azure_stack?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
-[![Translate](https://img.shields.io/badge/translate-zanata-blue.svg)](https://translate.zanata.org/zanata/project/view/manageiq-providers-azure_stack)
 
-ManageIQ plugin for the AzureStack provider.
+ManageIQ plugin for the Azure Stack provider.
 
 ## Development
 
-See the section on pluggable providers in the [ManageIQ Developer Setup](http://manageiq.org/docs/guides/developer_setup)
+See the section on plugins in the [ManageIQ Developer Setup](http://manageiq.org/docs/guides/developer_setup/plugins)
 
 For quick local setup run `bin/setup`, which will clone the core ManageIQ repository under the *spec* directory and setup necessary config files. If you have already cloned it, you can run `bin/update` to bring the core ManageIQ code up to date.
