+++ Rakefile
@@ -3,7 +3,7 @@ require 'bundler/setup'
 begin
   require 'rspec/core/rake_task'
 
-  APP_RAKEFILE = File.expand_path("../spec/manageiq/Rakefile", __FILE__)
+  APP_RAKEFILE = File.expand_path("spec/manageiq/Rakefile", __dir__)
   load 'rails/tasks/engine.rake'
   load 'rails/tasks/statistics.rake'
 rescue LoadError
