+++ lib/tasks_private/spec.rake
@@ -4,7 +4,6 @@ namespace :spec do
 end
 
 desc "Run all specs"
-RSpec::Core::RakeTask.new(:spec => ["app:test:initialize", "app:evm:compile_sti_loader", "app:test:providers_common"]) do |t|
-  spec_dir = File.expand_path("../../spec", __dir__)
-  EvmTestHelper.init_rspec_task(t, ['--require', File.join(spec_dir, 'spec_helper')])
+RSpec::Core::RakeTask.new(:spec => ['app:test:spec_deps', 'app:test:providers_common']) do |t|
+  EvmTestHelper.init_rspec_task(t)
 end
